Refactored removing habit, moved functionality to SharedHelper.java class
